No longer raise error when importing the '_cli' version of a module.
Otherwise the error is raised when using 'multiprocessing' (which imports the '_cli' module). The '_cli' modules are meant for running as scripts, but if user does 'import' them then they'll no longer get an error but will only have access to a 'main()' function anyway, so they'll find out soon enough that they imported the wrong module.